# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PG Workspaces core lifecycle API
  version: 0.3.0
  description: Proposed contract accompanying RFC-0001. Not a deployed service. Source/policy
    administration and pagination will be extended separately. All cross-resource
    lookups require current tenant/project authorization. No public server URL is
    specified.
security:
- bearerAuth: []
paths:
  /v1/projects/{project_id}/sources:
    post:
      operationId: createSource
      summary: Register an administrator-approved source; no arbitrary agent DSN
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceAccepted'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSource'
  /v1/projects/{project_id}/sources/{source_id}:
    get:
      operationId: getSource
      summary: Administrator source status and current epoch, including failed reseeds
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/source_id'
      responses:
        '200':
          description: Current source admission status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        default:
          description: Structured failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/sources/{source_id}/actions:
    post:
      operationId: reseedSource
      summary: Administrator reseed into a new immutable baseline generation
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/source_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceAction'
      responses:
        '202':
          description: New epoch and generation reserved; existing workspaces retain their lineage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceReseedAccepted'
        default:
          description: Structured failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/sources/{source_id}/barriers:
    post:
      operationId: issueBarrier
      summary: Issue a durable lower-bound source token after caller commit
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/source_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Barrier'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BarrierRequest'
  /v1/projects/{project_id}/baselines:
    get:
      operationId: listBaselines
      summary: List only baselines eligible under current principal/privacy policy
      parameters:
      - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineList'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/workspaces:
    post:
      operationId: createWorkspace
      summary: Create an isolated task workspace asynchronously
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAccepted'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspace'
  /v1/projects/{project_id}/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      summary: Read status and safe endpoint metadata
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/workspace_id'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteWorkspace
      summary: Revoke access and delete an expected generation asynchronously
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/workspace_id'
      - $ref: '#/components/parameters/ExpectedGeneration'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/workspaces/{workspace_id}/actions:
    post:
      operationId: workspaceAction
      summary: Pause, resume, reset, or extend absolute expiry
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/workspace_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
  /v1/projects/{project_id}/workspaces/{workspace_id}/credentials:
    post:
      operationId: issueCredential
      summary: Issue a generation-scoped credential after current authorization
      parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/workspace_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured failure; detail must not expose secrets or source
            row values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Service identity token with scoped tenant/project authority; not
        a database credential.
  parameters:
    project_id:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    source_id:
      name: source_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Bound to authenticated tenant/project, HTTP method/path, and normalized
        request hash. Different body under the same key returns 409.
...
  schemas:
    Source:
      type: object
      additionalProperties: false
      required: [id, connector, source_epoch, generation, status]
      properties:
        id:
          type: string
          format: uuid
        connector:
          type: string
          enum: [physical, logical]
        source_epoch:
          type: integer
          minimum: 1
        generation:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [registered, validating, seeding, streaming, blocked, unreachable, disabled]
        observed_at:
          type: string
          format: date-time
    SourceAction:
      type: object
      additionalProperties: false
      required: [action, expected_source_epoch]
      properties:
        action:
          const: reseed
        expected_source_epoch:
          type: integer
          minimum: 1
          maximum: 999999999
    SourceReseedAccepted:
      type: object
      additionalProperties: false
      required: [source_id, source_epoch, baseline_id, generation, operation]
      properties:
        source_id:
          type: string
          format: uuid
        source_epoch:
          type: integer
          minimum: 2
        baseline_id:
          type: string
          format: uuid
        generation:
          type: integer
          minimum: 2
        operation:
          $ref: '#/components/schemas/Operation'
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        operation_id:
          type: string
          format: uuid
        request_id:
          type: string
      required:
      - code
      - message
      - retryable
...
    CreateSource:
      type: object
      additionalProperties: false
      properties:
        connector:
          type: string
          enum:
          - physical
          - logical
        approved_endpoint_reference:
          type: string
        secret_reference:
          type: string
      required:
      - connector
      - approved_endpoint_reference
      - secret_reference
    SourceAccepted:
      type: object
      additionalProperties: false
      properties:
        source_id:
          type: string
          format: uuid
        operation:
          $ref: '#/components/schemas/Operation'
      required:
      - source_id
      - operation
    BarrierRequest:
      type: object
      additionalProperties: false
      properties:
        after_commit_asserted:
          const: true
      required:
      - after_commit_asserted
    Barrier:
      type: object
      additionalProperties: false
      properties:
        source_id:
          type: string
          format: uuid
        source_epoch:
          type: integer
          minimum: 1
        barrier_token:
          type: string
        issued_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
      required:
      - source_id
      - source_epoch
      - barrier_token
      - issued_at
      - expires_at